+++ ray-curator/ray_curator/stages/download/text/common_crawl/commoncrawl.py
@@ -15,7 +15,6 @@
 
 import os
 import subprocess
-import unicodedata
 import warnings
 from abc import ABC, abstractmethod
 from collections.abc import Iterator
@@ -26,8 +25,6 @@ from urllib.parse import urlparse
 
 import justext
 import lxml
-import pycld2 as cld2
-from charset_normalizer import detect
 from nemo_curator.datasets import DocumentDataset
 from nemo_curator.download.doc_builder import (
     DocumentDownloader,
@@ -43,37 +40,6 @@ from trafilatura.settings import DEFAULT_CONFIG as TRAFILATURA_DEFAULT_CONFIG
 from warcio.archiveiterator import ArchiveIterator
 
 NON_SPACED_LANGUAGES = ["THAI", "CHINESE", "JAPANESE", "KOREAN"]
-
-
-def decode_html(html_bytes: bytes) -> str | None:
-    # Convert from bytes to text using utf-8 encoding
-    try:
-        return html_bytes.decode("utf-8")
-    except UnicodeDecodeError:
-        # If utf-8 fails, try to find a different encoding
-        return try_decode_with_detected_encoding(html_bytes)
-
-
-def try_decode_with_detected_encoding(html_bytes: bytes) -> str | None:
-    detected_encoding = detect(html_bytes)["encoding"]
-    bad_detection = not detected_encoding or detected_encoding == "utf-8"
-    if bad_detection:
-        return None
-    try:
-        return html_bytes.decode(detected_encoding)
-    except:  # noqa: E722
-        return None
-
-
-def lang_detect(decoded_html: str) -> str:
-    try:
-        details = cld2.detect(decoded_html)[2]
-    except Exception:  # noqa: BLE001
-        # Remove control characters
-        cleaned_html = "".join(i for i in decoded_html if unicodedata.category(i)[0] != "C")
-        details = cld2.detect(cleaned_html)[2]
-
-    return details[0][0].upper()
 
 
 class HTMLExtractorAlgorithm(ABC):
